refactor(domain): 모든 상수를 관련있는 도메인 내부로 이동
