+++ src/main/java/domain/card/CardBoard.java
@@ -1,13 +1,15 @@
 package domain.card;
 
-import static domain.constant.GameRule.ACE_BONUS_SCORE;
-import static domain.constant.GameRule.BLACKJACK_CRITERION;
+import static domain.Game.INIT_CARD_COUNT;
 
 import domain.enums.Rank;
 import java.util.ArrayList;
 import java.util.List;
 
 public class CardBoard {
+    private static final int BLACKJACK_CRITERION = 21;
+    private static final int ACE_BONUS_SCORE = 10;
+
     private final List<Card> cards = new ArrayList<>();
 
     public void add(Card card) {
@@ -42,6 +44,17 @@ public class CardBoard {
 
     public boolean isBust() {
         return calculateScore() > BLACKJACK_CRITERION;
+    }
+
+    public boolean checkScoreUnderCriterion() {
+        return calculateScore() <= BLACKJACK_CRITERION;
+    }
+
+    public boolean isBlackjack() {
+        if (cards.size() != INIT_CARD_COUNT) {
+            return false;
+        }
+        return calculateScore() == BLACKJACK_CRITERION;
     }
 
     public List<Card> getCards() {
