--- src/main/java/domain/constant/GameRule.java
@@ -1,21 +1,0 @@
-package domain.constant;
-
-public class GameRule {
-    public static final int BLACKJACK_CRITERION = 21;
-    public static final int DEALER_HIT_CRITERION = 17;
-
-    public static final int ACE_BONUS_SCORE = 10;
-
-    public static final int MIN_PLAYER_NUMBER = 1;
-    public static final int MAX_PLAYER_NUMBER = 7;
-
-    public static final int MIN_NAME_LENGTH = 1;
-    public static final int MAX_NAME_LENGTH = 5;
-
-    public static final int INIT_CARD_COUNT = 2;
-
-    public static final int MIN_BET_AMOUNT = 1_000;
-    public static final int MAX_BET_AMOUNT = 300_000;
-
-    public static final int BET_AMOUNT_UNIT = 1_000;
-}
