+++ src/main/java/domain/participant/Name.java
@@ -1,10 +1,11 @@
 package domain.participant;
 
-import static domain.constant.GameRule.MAX_NAME_LENGTH;
-import static domain.constant.GameRule.MIN_NAME_LENGTH;
 import static message.ErrorMessage.PLAYER_NAME_OUT_OF_RANGE;
 
 public record Name(String name) {
+    public static final int MIN_NAME_LENGTH = 1;
+    public static final int MAX_NAME_LENGTH = 5;
+
     public Name {
         validateNameLength(name);
     }
